requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
banks==2.1.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.115.12
filetype==1.2.0
Flask==2.0.3
fonttools==4.58.4
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.2
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
kiwisolver==1.4.8
llama-cloud==0.1.19
llama-cloud-services==0.6.22
llama-index==0.12.35
llama-index-agent-openai==0.4.7
llama-index-cli==0.4.1
llama-index-core==0.12.35
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.38
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.22
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
openai==1.78.1
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.5.0
propcache==0.3.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.5.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
ruff==0.11.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uv==0.7.3
uvicorn==0.34.2
Werkzeug==2.2.2
wrapt==1.17.2
yarl==1.20.0
